requirements.txt
colorama
pynput
python-dotenv
schedule
datetime
